requirements-dev.txt
-r requirements.txt
pytest==5.3.2
pytest-asyncio==0.10.0
pytest-aiohttp==0.3.0
